+++ backend/logs/info/.2436d524048517d05bb330dfbac036842bb1c056-audit.json
@@ -29,6 +29,11 @@
             "date": 1688620278389,
             "name": "logs/info/2023-07-06.log",
             "hash": "f639e2d8e5ae6371fdaa4c1b54744f839f71f85d4a470f9e27e392b38bc49fe6"
+        },
+        {
+            "date": 1688960954435,
+            "name": "logs/info/2023-07-10.log",
+            "hash": "77d653fb5d1c50214206114f0f4d7340512c77cca35aee18b7fa98c265f0ef07"
         }
     ],
     "hashType": "sha256"
