+++ backend/logs/info/2023-07-06.log
@@ -5,3 +5,5 @@
 2023-07-06 14:15:02 info: GET /
 2023-07-06 14:15:08 info: GET /
 2023-07-06 14:15:32 info: The server is listening to 5000
+2023-07-06 15:25:01 info: The server is listening to 5000
+2023-07-06 17:33:59 info: The server is listening to 5000
