+++ backend/src/routes/index.ts
@@ -1,6 +1,6 @@
 import express from 'express';
 import swaggerUi from 'swagger-ui-express';
-import { logger } from '../config/winston';
+import logger from '../config/winston';
 import * as swaggerDocument from '../../swagger.json';
 
 const router = express.Router();
